+++ projects/_03_independent_cqc/_03_impute/tests/jobs/test_impute_ind_cqc_ascwds_and_pir.py
@@ -41,9 +41,7 @@ class ImputeIndCqcAscwdsAndPirTests(unittest.TestCase):
 
 class MainTests(ImputeIndCqcAscwdsAndPirTests):
     @patch(f"{PATCH_PATH}.utils.write_to_parquet")
-    @patch(
-        f"{PATCH_PATH}.convert_care_home_ratios_to_filled_posts_and_merge_with_filled_post_values"
-    )
+    @patch(f"{PATCH_PATH}.convert_care_home_ratios_to_posts")
     @patch(f"{PATCH_PATH}.cUtils.create_banded_bed_count_column")
     @patch(f"{PATCH_PATH}.model_calculate_rolling_average")
     @patch(f"{PATCH_PATH}.model_imputation_with_extrapolation_and_interpolation")
@@ -64,7 +62,7 @@ class MainTests(ImputeIndCqcAscwdsAndPirTests):
         model_imputation_with_extrapolation_and_interpolation_mock: Mock,
         model_calculate_rolling_average_mock: Mock,
         create_banded_bed_count_column_mock: Mock,
-        convert_care_home_ratios_to_filled_posts_and_merge_with_filled_post_values_mock: Mock,
+        convert_care_home_ratios_to_posts_mock: Mock,
         write_to_parquet_patch: Mock,
     ):
         read_from_parquet_patch.return_value = self.test_cleaned_ind_cqc_df
@@ -90,7 +88,7 @@ class MainTests(ImputeIndCqcAscwdsAndPirTests):
         )
         self.assertEqual(model_calculate_rolling_average_mock.call_count, 2)
         create_banded_bed_count_column_mock.assert_called_once()
-        convert_care_home_ratios_to_filled_posts_and_merge_with_filled_post_values_mock.assert_called_once()
+        convert_care_home_ratios_to_posts_mock.assert_called_once()
         write_to_parquet_patch.assert_called_once_with(
             ANY,
             self.ESTIMATES_DESTINATION,