+++ projects/_03_independent_cqc/_06_estimate_filled_posts/utils/models/non_res_with_dormancy.py
@@ -1,9 +1,6 @@
 from pyspark.ml.regression import LinearRegressionModel
 from pyspark.sql import DataFrame
 
-from projects._03_independent_cqc._06_estimate_filled_posts.utils.ml_model_metrics import (
-    save_model_metrics,
-)
 from projects._03_independent_cqc._06_estimate_filled_posts.utils.models.utils import (
     insert_predictions_into_pipeline,
     set_min_value,
@@ -15,14 +12,13 @@ def model_non_res_with_dormancy(
     locations_df: DataFrame,
     features_df: DataFrame,
     model_source: str,
-    metrics_destination: str,
 ) -> DataFrame:
     """
     Runs the non residential with dormancy model.
 
     This function loads and runs the non residential with dormancy
     predictions model on non residential with dormancy features data.
-    It then saves the model metrics and adds the predictions into the main dataset.
+    It then adds the predictions into the main dataset.
 
     Predictions are capped at a minimum of one.
 
@@ -31,7 +27,6 @@ def model_non_res_with_dormancy(
         features_df (DataFrame): A dataframe containing model features for the non res with dormancy model.
             This should only contain rows where dormancy is not null and the primary service type is non-residential.
         model_source (str): The file path to the non residential with dormancy model.
-        metrics_destination (str): The file path to the destination for saving metrics.
 
     Returns:
         DataFrame: A dataframe with non residential with dormancy model estimates added.
@@ -41,13 +36,6 @@ def model_non_res_with_dormancy(
     predictions_df = trained_model.transform(features_df)
 
     predictions_df = set_min_value(predictions_df, IndCqc.prediction, 1.0)
-
-    save_model_metrics(
-        predictions_df,
-        IndCqc.ascwds_pir_merged,
-        model_source,
-        metrics_destination,
-    )
 
     locations_df = insert_predictions_into_pipeline(
         locations_df,